move ws endpoint to routes/raft mod
