+++ metadata-server/src/routes/mod.rs
@@ -1,1 +1,2 @@
 pub mod api;
+pub mod raft;
